requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1
django-ninja==1.3.0
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
